modification de ft_strlen pour renvoyer un size_t

+++ fonc_libc/ft_strlen.c
@@ -10,7 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char	*s)
+#include <stdlib.h>
+
+size_t	ft_strlen(const char	*s)
 {
 	int	size;
